moved appowner tests one folder up
